Minor updates, SSL ignore and cookie parameters
Added 2 new constructor arguments. An argument to ignore ssl warnings,
set it to true to ignore ssl certificate errors. Also added a cookie
argument that makes it possible to add cookies to the requests send by
the client. There is also a new addCookie function that does the same,
but only for 1 cookie at a time, and they will not be used when loading
a description.
Commit also includes a bugfix in handling 100-continue headers.
Example updated to show new features in constructor.